+++ CHANGELOG.md
@@ -6,8 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
 - Added `nextTokenSuggestionsMulti` function to `optionparser_v2`.
 - Added `CHANGELOG.md` file.
+
+### Changed
+- Changed name of `PositionalIdentifier` to `Command` in `optionparser_v2`.
 
 
 ## [v0.1.0] - 2023-8-13